+++ booklog/src/App.tsx
@@ -1,12 +1,17 @@
 import { Header } from "./components/Header/Header";
 import { Body } from "./components/Body/Body";
-import { BooksContextProvider } from "./context";
+import { MyBooksProvider } from "./context/MyBooksProvider";
+import { BooksProvider } from "./context/BooksProvider";
 
 export const App = (): JSX.Element => {
   return (
-    <BooksContextProvider>
+    <>
       <Header />
-      <Body />
-    </BooksContextProvider>
+      <BooksProvider>
+        <MyBooksProvider>
+          <Body />
+        </MyBooksProvider>
+      </BooksProvider>
+    </>
   );
 };
