+++ booklog/src/types/index.ts
@@ -21,10 +21,13 @@ export type BookItem = {
 };
 
 export type MyBooksContextType = {
-  total: number;
-  bookItems: BooksResult["items"];
   myBooks: BookItem[];
-  fetchBooksApi: (query: string) => void;
   addMyBooks: (selectedBook: BookItem) => void;
   deleteMyBooks: (id: string) => void;
+};
+
+export type BooksContextType = {
+  total: number;
+  bookItems: BooksResult["items"];
+  fetchBooksApi: (query: string) => void;
 };
